composition mechanism
the `include()` function executes a script file in the same environment,
this allows the app creator to factor common code and variables out from
the app definition and into reusable submodules that are used to compose
the final app.

in the "basic apps" example, the common 'basic.lua' file defines
the `inputi` and `outputi` tables and the `reindex()` function.
Including this file from each app that needs it reduces code repetition.